Remove default link internal filters.
AppMeasurement send a beacon each time it detects an external link is
clicked.

When `linkExternalFilters` has no value, AppMeasurement uses
`linkInternalFilters` to detect the external links. If
`linkInternalFilters` is empty, then AppMeasurement sets
`linkInternalFilters` to window.location.hostname.

If we have `'javascript:', 'tel:', 'mailto:'`, then any link will be
considered as an external link (even if it's a link to a page on the
same domain).

One solution would be to append the hostname to linkInternalFilters at
runtime (inside the lib). But this automation might cause problems in
case the customer sets `linkInternalFilters` from custom code for
example.

In `search_discovery`, `linkInternalFilters` is empty by default, and we
decided to use the same approach here.