FEAT: flush queue on extension load (#543)
When loading in a Tags script with an instance of the same name, if existing instance has the same name, its will have its command queue flushed by the new instance. This allows users to create a base code instance and send it commands while the Tags script loads asynchronously.